requirements.txt
absl-py==0.15.0
addict==2.4.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==1.3.1
alembic==1.10.3
altair==4.2.2
altgraph==0.17.3
aniso8601==9.0.1
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
attrdict==2.0.1
attrs==22.1.0
Automat==22.10.0
Babel==2.12.1
backcall==0.2.0
bce-python-sdk==0.8.79
beautifulsoup4==4.11.2
black==21.6b0
bleach==5.0.1
Brotli==1.0.9
brotlipy==0.7.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.4
comm @ file:///croot/comm_1671231121260/work
constantly==15.1.0
contourpy==1.0.5
cryptography==39.0.2
cssselect==1.2.0
cssutils==2.6.0
cycler==0.11.0
Cython==0.29.33
dataclasses==0.6
debugpy==1.5.1
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml==0.7.1
dnspython==2.3.0
docker-pycreds==0.4.0
efficientnet-pytorch==0.7.1
einops==0.2.0
email-validator==1.1.3
engineering-notation==0.8.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.2
fastapi==0.92.0
fastjsonschema==2.16.2
ffmpy==0.3.0
filelock==3.8.0
fire==0.5.0
Flask==2.0.1
flask-babel==3.0.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
flatbuffers==1.12
fonttools==4.29.1
frozenlist==1.3.3
fsspec==2023.3.0
gast==0.3.3
gdown==4.4.0
gevent==22.10.2
geventhttpclient==2.0.2
gitdb==4.0.10
GitPython==3.1.30
google-api-core==2.11.0
google-api-python-client==2.81.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.58.0
gradio==3.20.0
greenlet==2.0.2
grpcio==1.32.0
gunicorn==20.1.0
h11==0.14.0
h5py==2.10.0
html-parser==0.2
html5lib==1.1
httpcore==0.16.3
httplib2==0.21.0
httpx==0.23.3
huggingface-hub==0.14.1
hyperlink==21.0.0
idna==3.3
imageio @ file:///home/conda/feedstock_root/build_artifacts/imageio_1686552404404/work
imgaug==0.4.0
importlib-metadata==4.11.1
importlib-resources==5.9.0
imutils==0.5.4
incremental==22.10.0
install==1.3.5
ipykernel==5.5.6
ipython==8.0.1
ipython-genutils==0.2.0
ipywidgets==8.0.2
itemadapter==0.7.0
itemloaders==1.0.6
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.0.1
jmespath==1.0.1
joblib==1.1.0
json5 @ file:///tmp/build/80754af9/json5_1624432770122/work
jsonschema==4.16.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.4
jupyter-core==4.9.2
jupyter-server @ file:///croot/jupyter_server_1671707632269/work
jupyterlab @ file:///croot/jupyterlab_1675354114448/work
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
jupyterlab_server @ file:///croot/jupyterlab_server_1672127357747/work
keras==2.13.1
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Levenshtein==0.20.9
libclang==15.0.6.1
linkify-it-py==2.0.0
lmdb==1.3.0
lxml==4.9.1
Markdown==3.3.6
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mdit-py-plugins==0.3.3
mdurl==0.1.2
MedPy==0.4.0
mistune==2.0.4
mkl-fft==1.3.6
mkl-random @ file:///work/mkl/mkl_random_1682950433854/work
mkl-service==2.4.0
mmcv==2.0.1
mmengine==0.8.2
mpmath==1.2.1
multidict==6.0.4
munch==2.5.0
munkres==1.1.4
mypy-extensions==0.4.3
nbclassic @ file:///croot/nbclassic_1668174957779/work
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.6.1
nest-asyncio==1.5.4
networkx==2.6.3
nltk==3.8.1
notebook==6.4.12
notebook_shim @ file:///croot/notebook-shim_1668160579331/work
numpy==1.19.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.0
onnx==1.13.1
opencv-contrib-python==4.6.0.66
opencv-python==4.5.5.64
opencv-python-headless==4.6.0.66
openpyxl==3.1.1
opt-einsum==3.3.0
orjson==3.8.7
packaging==21.3
paddle-bfloat==0.1.7
paddleocr==2.6.1.3
paddlepaddle-gpu==2.4.2.post117
pandas==1.4.1
pandocfilters==1.5.0
parsel==1.7.0
parso==0.8.3
pathspec==0.9.0
pathtools==0.1.2
pdf2docx==0.5.6
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pkgutil_resolve_name==1.3.10
platformdirs==2.5.1
ply==3.11
prefetch-generator==1.0.1
premailer==3.10.0
pretrainedmodels==0.7.4
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.28
Protego==0.2.1
protobuf==3.19.4
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
py-translate==1.0.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post4
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.5
PyDispatcher==2.0.7
pydub==0.25.1
Pygments==2.11.2
pyinstaller==5.11.0
pyinstaller-hooks-contrib==2023.3
pymongo==4.3.3
PyMuPDF==1.20.2
pyOpenSSL==23.0.0
pyparsing==3.0.7
PyQt5-sip==12.11.0
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.4
python-docx==0.8.11
python-multipart==0.0.6
python-rapidjson==1.9
python-version==0.0.2
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0.1
pyzmq==22.3.0
qtconsole==5.3.2
QtPy==2.2.0
qudida==0.0.4
queuelib==1.6.2
rapidfuzz==2.13.7
rarfile==4.0
regex==2022.1.18
requests==2.27.1
requests-file==1.5.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.4.2
rsa==4.8
scapy==2.5.0
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
Scrapy==2.8.0
seaborn==0.12.2
segmentation-models-pytorch==0.3.2
Send2Trash==1.8.0
sentry-sdk==1.12.1
service-identity==21.1.0
setproctitle==1.3.2
Shapely==1.8.4
shortuuid==1.0.11
SimpleITK==2.2.1
sip @ file:///tmp/abs_44cd77b_pu/croots/recipe/sip_1659012365470/work
six==1.15.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.23
stack-data==0.2.0
starlette==0.25.0
sympy==1.11.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.4.4
tensorflow-estimator==2.4.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==1.1.0
terminado==0.15.0
Theano==1.0.5
threadpoolctl==3.1.0
tifffile==2022.2.9
timm==0.5.4
tinycss2==1.1.1
tk==0.1.0
tk-tools==0.16.0
tkintertable==1.3.3
tldextract==3.4.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.10.1+cu111
torch-tb-profiler==0.4.1
torchaudio==0.10.1+cu111
torchvision==0.11.2+cu111
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
tritonclient==2.31.0
Twisted==22.10.0
typing_extensions==4.7.1
uc-micro-py==1.0.1
uritemplate==4.1.1
urllib3==1.26.8
uvicorn==0.20.0
vietocr==0.3.9
visualdl==2.5.1
w3lib==2.1.1
wandb==0.13.5
wcwidth==0.2.5
webencodings==0.5.1
websocket-client @ file:///tmp/build/80754af9/websocket-client_1614804261064/work
websockets==10.4
Werkzeug==2.0.3
widgetsnbextension==4.0.3
wrapt==1.12.1
WTForms==2.3.3
x2paddle==1.4.0
xlrd==2.0.1
yapf==0.40.1
yarl==1.8.2
yellowbrick==1.5
zipp==3.7.0
zope.event==4.6
zope.interface==5.5.2